:truck: Move entity to shared
