+++ src/LearningObjects/Hierarchy/HierarchyInteractor.ts
@@ -1,6 +1,6 @@
 import { DataStore, ParentLearningObjectQuery } from '../../shared/interfaces/DataStore';
 import { UserToken } from '../../shared/types';
-import { LearningObject } from '../../entity';
+import { LearningObject } from '../../shared/entity';
 import { toArray, hasReadAccessByCollection, LearningObjectState, handleError } from '../../interactors/LearningObjectInteractor';
 import { ResourceError, ResourceErrorReason } from '../../errors';
