+++ src/LearningObjects/LearningObjectRouteHandler.ts
@@ -4,7 +4,7 @@ import { DataStore } from '../shared/interfaces/DataStore';
 import { FileManager, LibraryCommunicator } from '../shared/interfaces/interfaces';
 import { UserToken } from '../shared/types';
 import * as LearningObjectInteractor from './LearningObjectInteractor';
-import { LearningObject } from '../entity';
+import { LearningObject } from '../shared/entity';
 import { FileMeta } from './typings';
 
 /**
