+++ src/LearningObjects/PDFKitDriver.ts
@@ -3,7 +3,7 @@ import { LEARNING_OBJECT_ROUTES } from '../routes';
 import { FileManager } from '../shared/interfaces/interfaces';
 import { FileUpload } from '../shared/interfaces/FileManager';
 import * as striptags from 'striptags';
-import { LearningObject, LearningOutcome } from '../entity';
+import { LearningObject, LearningOutcome } from '../shared/entity';
 
 type GradientVector = [number, number, number, number];
 type PDFHeaderAlignment = 'left' | 'right' | 'center' | 'justify';
