+++ src/shared/entity/errors/entity-error.ts
