+++ src/shared/entity/index.ts
