+++ src/shared/entity/learning-object/learning-object.ts
