+++ src/shared/entity/learning-object/submittable-learning-object.spec.ts
