+++ src/shared/entity/learning-outcome/submittable-learning-outcome.spec.ts
