+++ src/shared/entity/learning-outcome/submittable-learning-outcome.ts
