+++ src/shared/entity/outcome/outcome.ts
