+++ src/shared/entity/standard-outcome/standard-outcome.spec.ts
