+++ src/shared/entity/standard-outcome/standard-outcome.ts
