+++ src/shared/entity/user/user.spec.ts
