+++ src/shared/interfaces/LibraryCommunicator.ts
@@ -1,4 +1,4 @@
-import { LearningObject } from '../../entity';
+import { LearningObject } from '../entity';
 
 export interface LibraryCommunicator {
   getMetrics(objectID: string): Promise<LearningObject.Metrics>;
