+++ src/tests/mock-drivers/MockLibraryDriver.ts
@@ -1,6 +1,6 @@
 import { LibraryCommunicator } from '../../shared/interfaces/LibraryCommunicator';
 import { MOCK_OBJECTS } from '../mocks';
-import { LearningObject } from '../../entity';
+import { LearningObject } from '../../shared/entity';
 
 export class MockLibraryDriver implements LibraryCommunicator {
   getMetrics(objectID: string): Promise<LearningObject.Metrics> {
